requirements.txt
flask
easyocr
transformers
torch
pillow
numpy
